Add support for color inversion
Currently, when drawing to the screen, using "Color::WHITE" paints a black pixel
and "Color::BLACK" paints a white one. Also when using grayscale images, the image appears inverted (i.e. negative) by default.

With this commit, a new option "inverted" is added, to invert the colors and let "white" show as black and "black" as white when set to true.
In order to not break existing implementations, the default has been set to "True".